Use std::numeric_limits to avoid some GCC-15 problems.
FLT_SNAN and DBL_SNAN are nonstandard C++ macros which are now defined
by C23. Since SAMRAI checks for the presence of these macros with
cpp (instead of CXX) it erroneously, for compilers defaulting to C23,
believes these macros are available in C++.

Fix this by just using std::numeric_limits, which is always available.